move cp under the files namespace
