+++ packages/cli/src/commands/sandbox/index.ts
@@ -1,5 +1,5 @@
 import { defineNamespace } from "../../core/define-namespace.ts";
-import { sandboxCpCommand } from "./cp.ts";
+import { sandboxCpMovedCommand } from "./cp.ts";
 import { sandboxCreateCommand } from "./create.ts";
 import { sandboxDeleteCommand } from "./delete.ts";
 import { sandboxEnvNamespace } from "./env/index.ts";
@@ -17,10 +17,10 @@ export const sandboxNamespace = defineNamespace(
     sandboxListCommand,
     sandboxDeleteCommand,
     sandboxExecCommand,
-    sandboxCpCommand,
     sandboxFilesNamespace,
     sandboxSshCommand,
     sandboxUrlNamespace,
     sandboxEnvNamespace,
+    sandboxCpMovedCommand,
   ],
 );
